refactor: combine BasicMeta and Header component, consolidate meta components
