+++ config.json
@@ -1,7 +1,7 @@
 {
-  "base_url": "https://sdohplace.org/news",
-  "site_title": "News | SDOH & Place Project",
-  "site_description": "News and updates related to the SDOH & Place Project from Heathy Regions & Policies Lab at Univiersity of Illinois, Urbana-Champaign.",
+  "base_url": "https://sdohplace.org",
+  "site_title": "SDOH & Place Project",
+  "site_description": "Website for the SDOH & Place project from Healthy Regions & Policies Lab at Univiersity of Illinois, Urbana-Champaign.",
   "site_keywords": [
     {
       "keyword": "Public Health"
