+++ src/pages/news/page/[page].tsx
@@ -1,13 +1,12 @@
 import { GetStaticPaths, GetStaticProps } from "next";
-import Head from "next/head";
 import Layout from "@/components/Layout";
-import BasicMeta from "@/components/news/meta/BasicMeta";
-import OpenGraphMeta from "@/components/news/meta/OpenGraphMeta";
-import TwitterCardMeta from "@/components/news/meta/TwitterCardMeta";
+import Header from "@/components/meta/Header";
+import OpenGraphMeta from "@/components/meta/OpenGraphMeta";
+import TwitterCardMeta from "@/components/meta/TwitterCardMeta";
 import PostList from "@/components/news/PostList";
-import config from "../../../lib/config";
-import { countPosts, listPostContent, PostContent } from "../../../lib/posts";
-import { listTags, TagContent } from "../../../lib/tags";
+import config from "@/lib/config";
+import { countPosts, listPostContent, PostContent } from "@/lib/posts";
+import { listTags, TagContent } from "@/lib/tags";
 
 type Props = {
   posts: PostContent[];
@@ -23,7 +22,7 @@ export default function Page({ posts, tags, pagination, page }: Props) {
   const title = "All posts";
   return (
     <Layout>
-      <BasicMeta url={url} title={title} />
+      <Header url={url} title={title} />
       <OpenGraphMeta url={url} title={title} />
       <TwitterCardMeta url={url} title={title} />
       <PostList posts={posts} tags={tags} pagination={pagination} />
