+++ src/pages/news/tags/[[...slug]].tsx
@@ -1,12 +1,12 @@
 import { GetStaticPaths, GetStaticProps } from "next";
 import Layout from "@/components/Layout";
-import BasicMeta from "@/components/news/meta/BasicMeta";
-import OpenGraphMeta from "@/components/news/meta/OpenGraphMeta";
-import TwitterCardMeta from "@/components/news/meta/TwitterCardMeta";
+import Header from "@/components/meta/Header";
+import OpenGraphMeta from "@/components/meta/OpenGraphMeta";
+import TwitterCardMeta from "@/components/meta/TwitterCardMeta";
 import TagPostList from "@/components/news/TagPostList";
-import config from "../../../lib/config";
-import { countPosts, listPostContent, PostContent } from "../../../lib/posts";
-import { getTag, listTags, TagContent } from "../../../lib/tags";
+import config from "@/lib/config";
+import { countPosts, listPostContent, PostContent } from "@/lib/posts";
+import { getTag, listTags, TagContent } from "@/lib/tags";
 
 type Props = {
   posts: PostContent[];
@@ -22,7 +22,7 @@ export default function Index({ posts, tag, pagination, page }: Props) {
   const title = tag.name;
   return (
     <Layout>
-      <BasicMeta url={url} title={title} />
+      <Header url={url} title={title} />
       <OpenGraphMeta url={url} title={title} />
       <TwitterCardMeta url={url} title={title} />
       <TagPostList posts={posts} tag={tag} pagination={pagination} />
