+++ src/pages/showcase/index.tsx
@@ -1,8 +1,8 @@
 import { GetStaticProps } from "next";
 import Layout from "@/components/Layout";
-import BasicMeta from "@/components/news/meta/BasicMeta";
-import OpenGraphMeta from "@/components/news/meta/OpenGraphMeta";
-import TwitterCardMeta from "@/components/news/meta/TwitterCardMeta";
+import Header from "@/components/meta/Header";
+import OpenGraphMeta from "@/components/meta/OpenGraphMeta";
+import TwitterCardMeta from "@/components/meta/TwitterCardMeta";
 import ShowcaseList from "@/components/showcase/ShowcaseList";
 import config from "../../lib/config";
 import {
@@ -11,7 +11,6 @@ import {
   ShowcaseContent
 } from "../../lib/showcases";
 import { listTags, TagContent } from "../../lib/tags";
-import Head from "next/head";
 
 type Props = {
   posts: ShowcaseContent[];
@@ -26,7 +25,7 @@ export default function Index({ posts, tags, pagination }: Props) {
   const title = "Showcase";
   return (
     <Layout page_header={"Fellows Showcase"}>
-      <BasicMeta url={url} title={title} />
+      <Header url={url} title={title} />
       <OpenGraphMeta url={url} title={title} />
       <TwitterCardMeta url={url} title={title} />
       <ShowcaseList posts={posts} pagination={pagination} />
